Add repositories and view model for company management; implement AsyncResult for repository methods, create CompaniesViewModel for state management, and introduce new UI components for asset tree visualization.
