+++ lib/core/domain/company_location/company_location_repository.dart
@@ -1,6 +1,7 @@
+import '../result.dart';
 import '../uid.dart';
 import 'company_location.dart';
 
 abstract class CompanyLocationRepository {
-  Future<List<CompanyLocation>> fromCompany(Uid companyId);
+  AsyncResult<List<CompanyLocation>> fromCompany(Uid companyId);
 }
