+++ lib/infra/repositories/tractian_company_asset_repository.dart
@@ -1,5 +1,6 @@
 import '../../core/domain/company_asset/company_asset.dart';
 import '../../core/domain/company_asset/company_asset_repository.dart';
+import '../../core/domain/result.dart';
 import '../../core/domain/uid.dart';
 import '../services/http/tractian_http_client.dart';
 
@@ -11,8 +12,8 @@ class TractianCompanyAssetRepository implements CompanyAssetRepository {
   }) : _httpClient = httpClient;
 
   @override
-  Future<List<CompanyAsset>> fromCompany(Uid companyId) async {
+  AsyncResult<List<CompanyAsset>> fromCompany(Uid companyId) async {
     final response = await _httpClient.get('/companies/$companyId/assets');
-    return response.companyAssets;
+    return Result.ok(response.companyAssets);
   }
 }
