+++ lib/infra/repositories/tractian_company_location_repository.dart
@@ -1,5 +1,6 @@
 import '../../core/domain/company_location/company_location.dart';
 import '../../core/domain/company_location/company_location_repository.dart';
+import '../../core/domain/result.dart';
 import '../../core/domain/uid.dart';
 import '../services/http/tractian_http_client.dart';
 
@@ -11,8 +12,8 @@ class TractianCompanyLocationRepository implements CompanyLocationRepository {
   }) : _httpClient = httpClient;
 
   @override
-  Future<List<CompanyLocation>> fromCompany(Uid companyId) async {
+  AsyncResult<List<CompanyLocation>> fromCompany(Uid companyId) async {
     final response = await _httpClient.get('/companies/$companyId/locations');
-    return response.companyLocations;
+    return Result.ok(response.companyLocations);
   }
 }
