+++ lib/infra/repositories/tractian_company_repository.dart
@@ -1,5 +1,6 @@
 import '../../core/domain/company/company.dart';
 import '../../core/domain/company/company_repository.dart';
+import '../../core/domain/result.dart';
 import '../services/http/tractian_http_client.dart';
 
 class TractianCompanyRepository implements CompanyRepository {
@@ -10,8 +11,8 @@ class TractianCompanyRepository implements CompanyRepository {
   }) : _httpClient = httpClient;
 
   @override
-  Future<List<Company>> get all async {
+  AsyncResult<List<Company>> get all async {
     final response = await _httpClient.get('/companies');
-    return response.companies;
+    return Result.ok(response.companies);
   }
 }
